+++ simpletransformers/classification/classification_model.py
@@ -1639,16 +1639,10 @@ class ClassificationModel:
 
             if not self.args.sliding_window:
                 # ROC`
-                wandb.log({"roc": wandb.plots.ROC(truth, model_outputs, labels_list)})
+                wandb.log({"roc": wandb.plot.roc_curve(truth, model_outputs, labels_list)})
 
                 # Precision Recall
-                wandb.log(
-                    {
-                        "pr": wandb.plots.precision_recall(
-                            truth, model_outputs, labels_list
-                        )
-                    }
-                )
+                wandb.log({"pr": wandb.plot.pr_curve(truth, model_outputs, labels_list)})
 
         return results, model_outputs, wrong
 
@@ -2241,19 +2235,26 @@ class ClassificationModel:
                     final_preds = []
                     for pred_row in preds:
                         mean_pred = np.mean(pred_row)
-                        print(mean_pred)
+                        # print(mean_pred)
                         final_preds.append(mean_pred)
-                    preds = np.array(final_preds)
                 else:
                     preds = [np.argmax(pred, axis=1) for pred in preds]
                     final_preds = []
                     for pred_row in preds:
-                        mode_pred, counts = mode(pred_row)
-                        if len(counts) > 1 and counts[0] == counts[1]:
+                        val_freqs_desc = Counter(pred_row).most_common()
+                        if (
+                            len(val_freqs_desc) > 1
+                            and val_freqs_desc[0][1] == val_freqs_desc[1][1]
+                        ):
                             final_preds.append(args.tie_value)
                         else:
-                            final_preds.append(mode_pred[0])
-                    preds = np.array(final_preds)
+                            final_preds.append(val_freqs_desc[0][0])
+                        # mode_pred, counts = mode(pred_row)
+                        # if len(counts) > 1 and counts[0] == counts[1]:
+                        #     final_preds.append(args.tie_value)
+                        # else:
+                        #     final_preds.append(mode_pred[0])
+                preds = np.array(final_preds)
             elif not multi_label and args.regression is True:
                 preds = np.squeeze(preds)
                 model_outputs = preds