+++ simpletransformers/seq2seq/seq2seq_model.py
@@ -36,7 +36,9 @@ from transformers import (
     BartTokenizerFast,
     MBartConfig,
     MBartForConditionalGeneration,
-    MBartTokenizer,
+    MBartTokenizerFast,
+    MBart50Tokenizer,
+    MBart50TokenizerFast,
     BertConfig,
     BertModel,
     BertTokenizerFast,
@@ -100,7 +102,8 @@ logger = logging.getLogger(__name__)
 MODEL_CLASSES = {
     "auto": (AutoConfig, AutoModel, AutoTokenizer),
     "bart": (BartConfig, BartForConditionalGeneration, BartTokenizerFast),
-    "mbart": (MBartConfig, MBartForConditionalGeneration, MBartTokenizer),
+    "mbart": (MBartConfig, MBartForConditionalGeneration, MBartTokenizerFast),
+    "mbart50": (MBartConfig, MBartForConditionalGeneration, MBart50TokenizerFast),
     "bert": (BertConfig, BertModel, BertTokenizerFast),
     "camembert": (CamembertConfig, CamembertModel, CamembertTokenizerFast),
     "distilbert": (DistilBertConfig, DistilBertModel, DistilBertTokenizerFast),
@@ -307,9 +310,9 @@ class Seq2SeqModel:
             else:
                 config_class, model_class, tokenizer_class = MODEL_CLASSES[encoder_type]
 
-            if encoder_decoder_type in ["bart", "mbart", "marian"]:
+            if encoder_decoder_type in ["bart", "mbart", "mbart50", "marian"]:
                 self.model = model_class.from_pretrained(encoder_decoder_name)
-                if encoder_decoder_type in ["bart", "mbart"]:
+                if encoder_decoder_type in ["bart", "mbart","mbart50"]:
                     self.encoder_tokenizer = tokenizer_class.from_pretrained(
                         encoder_decoder_name
                     )
@@ -1261,7 +1264,7 @@ class Seq2SeqModel:
                     return_tensors="pt",
                     truncation=True,
                 )["input_ids"]
-            elif self.args.model_type in ["mbart"]:
+            elif self.args.model_type in ["mbart", "mbart50"]:
                 input_ids = self.encoder_tokenizer.prepare_seq2seq_batch(
                     src_texts=batch,
                     max_length=self.args.max_seq_length,
@@ -1314,15 +1317,18 @@ class Seq2SeqModel:
                     num_return_sequences=self.args.num_return_sequences,
                 )
             elif self.args.model_type in ["mbart"]:
-                tgt_lang_token = self.decoder_tokenizer._convert_token_to_id(
-                    self.args.tgt_lang
-                )
+                
+                # tgt_lang_token = self.decoder_tokenizer._convert_token_to_id_with_added_voc(
+                #     self.args.tgt_lang
+                # )
 
                 outputs = self.model.generate(
                     input_ids=input_ids,
-                    decoder_start_token_id=tgt_lang_token,
+                    # decoder_start_token_id=tgt_lang_token,
+                    decoder_start_token_id=self.decoder_tokenizer.lang_code_to_id[self.args.tgt_lang],
                     num_beams=self.args.num_beams,
-                    max_length=self.args.max_length,
+                    # max_length=self.args.max_length,
+                    max_new_tokens=self.args.max_length,
                     length_penalty=self.args.length_penalty,
                     early_stopping=self.args.early_stopping,
                     repetition_penalty=self.args.repetition_penalty,
@@ -1330,6 +1336,28 @@ class Seq2SeqModel:
                     top_k=self.args.top_k,
                     top_p=self.args.top_p,
                     num_return_sequences=self.args.num_return_sequences,
+                )
+            elif self.args.model_type in ["mbart50"]:
+                
+                # tgt_lang_token = self.decoder_tokenizer._convert_token_to_id_with_added_voc(
+                #     self.args.tgt_lang
+                # )
+
+                outputs = self.model.generate(
+                    input_ids=input_ids,
+                    # decoder_start_token_id=tgt_lang_token,
+                    decoder_start_token_id=self.decoder_tokenizer.lang_code_to_id[self.args.tgt_lang],
+                    num_beams=self.args.num_beams,
+                    # max_length=self.args.max_length,
+                    max_new_tokens=self.args.max_length,
+                    length_penalty=self.args.length_penalty,
+                    early_stopping=self.args.early_stopping,
+                    repetition_penalty=self.args.repetition_penalty,
+                    do_sample=self.args.do_sample,
+                    top_k=self.args.top_k,
+                    top_p=self.args.top_p,
+                    num_return_sequences=self.args.num_return_sequences,
+                    forced_bos_token_id=self.decoder_tokenizer.lang_code_to_id[self.args.tgt_lang],
                 )
             elif self.args.model_type in ["rag-token", "rag-sequence"]:
                 outputs = self.model.generate(
@@ -1494,7 +1522,7 @@ class Seq2SeqModel:
                     encoder_tokenizer, decoder_tokenizer, args, data, mode
                 )
             else:
-                if args.model_type in ["bart", "mbart", "marian"]:
+                if args.model_type in ["bart", "mbart", "mbart50", "marian"]:
                     return SimpleSummarizationDataset(
                         encoder_tokenizer, self.args, data, mode
                     )
@@ -1543,6 +1571,7 @@ class Seq2SeqModel:
             if self.args.model_type in [
                 "bart",
                 "mbart",
+                "mbart50",
                 "marian",
                 "rag-token",
                 "rag-sequence",
@@ -1554,6 +1583,7 @@ class Seq2SeqModel:
                 if self.args.model_type in [
                     "bart",
                     "mbart",
+                    "mbart50",
                     "marian",
                     "rag-token",
                     "rag-sequence",
@@ -1630,7 +1660,7 @@ class Seq2SeqModel:
                 "attention_mask": source_mask.to(device),
                 "labels": labels.to(device),
             }
-        elif self.args.model_type in ["mbart"]:
+        elif self.args.model_type in ["mbart","mbart50"]:
             inputs = {
                 "input_ids": batch["input_ids"].to(device),
                 "attention_mask": batch["attention_mask"].to(device),