+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.64.0] - 2023-04-24
+
+### Changed
+
+- Updates for compatibility with transformers v4.31.0
+
+### Fixed
+
+- Fixed bug in `ClassificationModel` when using `sliding_window`
+
+## [0.63.11] - 2023-04-24
+
+### Fixed
+
+- Retrieval model bug fix
+
+## [0.63.10] - 2023-03-30
+
+### Added
+
+- mBART50 [whr778](https://github.com/whr778)
+
+### Fixed
+
+- Fixed wandb deprecated plots [luketudge](https://github.com/luketudge)
+
 ## [0.63.9] - 2022-09-18
 
 ### Added